refactor(CO-3211): update message type imports to use specific paths for better clarity
