+++ src/api/get-conv-soap-api.ts
@@ -9,13 +9,10 @@ import { map } from 'lodash';
 import { MAIL_VERIFICATION_HEADERS } from 'constants/index';
 import { normalizeConversations } from 'normalizations/normalize-conversation';
 import { normalizeMailMessageFromSoap } from 'normalizations/normalize-message';
-import type {
-	GetConvParameters,
-	GetConvRequest,
-	GetConvResponse,
-	IncompleteMessage,
-	NormalizedConversation
-} from 'types/index.d';
+import { NormalizedConversation } from 'types/conversations';
+import { IncompleteMessage } from 'types/messages';
+import { GetConvRequest, GetConvResponse } from 'types/soap/get-conv';
+import { GetConvParameters } from 'types/soap/soap';
 
 export const getConvSoapApi = async ({
 	conversationId,
