+++ src/commons/print-conversation/get-attachments.ts
@@ -7,7 +7,7 @@
 import { t } from '@zextras/carbonio-shell-ui';
 import { map } from 'lodash';
 
-import type { MailMessage } from 'types/index.d';
+import { MailMessage } from 'types/messages';
 
 export const getAttachments = ({ msg }: { msg: MailMessage }): string =>
 	`
