+++ src/hooks/actions/tests/use-msg-print.test.ts
@@ -9,10 +9,10 @@ import { FOLDERS } from '@zextras/carbonio-ui-commons';
 
 import { setupHook } from '@test-setup';
 import { createSoapAPIInterceptor } from '@test-utils/network/msw/create-api-interceptor';
+import { generateMessage } from '__test__/generators/generateMessage';
 import { FOLDERS_DESCRIPTORS } from 'constants/index';
 import { useMsgPrintDescriptor, useMsgPrintFn } from 'hooks/actions/use-msg-print';
-import { generateMessage } from '__test__/generators/generateMessage';
-import { MailMessage } from 'types/index.d';
+import { MailMessage } from 'types/messages';
 
 describe('useMsgPrintDescripto', () => {
 	const msg = generateMessage({ isComplete: true });
