+++ src/hooks/actions/tests/use-msg-show-original.test.ts
@@ -9,13 +9,13 @@ import { FOLDERS } from '@zextras/carbonio-ui-commons';
 
 import { setupHook } from '@test-setup';
 import { createSoapAPIInterceptor } from '@test-utils/network/msw/create-api-interceptor';
+import { generateMessage } from '__test__/generators/generateMessage';
 import { FOLDERS_DESCRIPTORS } from 'constants/index';
 import {
 	useMsgShowOriginalDescriptor,
 	useMsgShowOriginalFn
 } from 'hooks/actions/use-msg-show-original';
-import { generateMessage } from '__test__/generators/generateMessage';
-import { MailMessage } from 'types/index.d';
+import { MailMessage } from 'types/messages';
 
 describe('useMsgShowOriginal', () => {
 	const msg = generateMessage();
