+++ src/store/emails/actions/tests/get-message-with-existing-participants.test.ts
@@ -1,14 +1,16 @@
-import { ParticipantRole } from '@zextras/carbonio-ui-commons';
 /*
  * SPDX-FileCopyrightText: 2025 Zextras <https://www.zextras.com>
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
+import { ParticipantRole } from '@zextras/carbonio-ui-commons';
+
 import { generateCompleteMessageFromAPI } from '../../../../__test__/generators/api';
 import { createSoapAPIInterceptor } from '@test-utils/network/msw/create-api-interceptor';
 import { getMessageWithExistingParticipantsEmailStoreAction } from 'store/emails/actions/get-message-with-existing-participants';
-import { GetMsgRequest, GetMsgResponse, MailMessage } from 'types/index.d';
+import { MailMessage } from 'types/messages';
+import { GetMsgRequest, GetMsgResponse } from 'types/soap/get-msg';
 
 const stubGetMsgApi = (response: any): Promise<GetMsgRequest> =>
 	createSoapAPIInterceptor<GetMsgRequest, GetMsgResponse>('GetMsg', response);
