+++ src/store/utils.ts
@@ -6,7 +6,8 @@
 import { getUserAccount } from '@zextras/carbonio-shell-ui';
 import { includes, map } from 'lodash';
 
-import type { FetchConversationsReturn, MailMessage, NormalizedConversation } from 'types/index.d';
+import { FetchConversationsReturn, NormalizedConversation } from 'types/conversations';
+import { MailMessage } from 'types/messages';
 
 /**
  * Extracts all ids from conversations and messages
