+++ src/types/index.d.ts
@@ -7,5 +7,4 @@
 export * from '@zextras/carbonio-ui-commons';
 export * from './api';
 export * from './conversations';
-export * from './messages';
 export * from './soap';
