+++ src/views/app/detail-panel/conversation-message-preview.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 
 import { Padding } from '@zextras/carbonio-design-system';
 
-import { MailMessage } from 'types/index.d';
+import { MailMessage } from 'types/messages';
 import MailPreview from 'views/app/detail-panel/preview/mail-preview';
 
 export type ConversationMessagePreviewProps = {
