+++ src/views/app/detail-panel/eml-preview-panel-container.tsx
@@ -12,7 +12,7 @@ import type { EmlRouteParams } from '../../../types/routes';
 import { getMsgSoapApi } from 'api/get-msg-soap-api';
 import { isFocusModeMailView } from 'helpers/external-tabs';
 import { normalizeMailMessageFromSoap } from 'normalizations/normalize-message';
-import { MailMessage } from 'types/index.d';
+import { MailMessage } from 'types/messages';
 import { MessagePreviewPanel } from 'views/app/detail-panel/message-preview-panel';
 
 export const EmlPreviewPanelContainer = (): React.JSX.Element => {
