+++ src/views/app/detail-panel/preview/parts/info-block/mail-sensitivity-icon.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { Icon, Row, Tooltip } from '@zextras/carbonio-design-system';
 import { useTranslation } from 'react-i18next';
 
-import { Sensitivity } from 'types/index.d';
+import { Sensitivity } from 'types/messages';
 import {
 	getMailSensitivityIconColor,
 	getMailSensitivityLabel
