+++ src/views/search/list/message/tests/search-message-list-item-wrapper.test.tsx
@@ -11,7 +11,7 @@ import { render, screen, waitFor } from '@testing-library/react';
 import { setupTest } from '@test-setup';
 import { populateMessagesInEmailStore } from '__test__/generators/generateMessage';
 import * as storeModule from 'store/emails/store';
-import { MailMessage } from 'types/index.d';
+import { MailMessage } from 'types/messages';
 import { SearchMessageListItemWrapper } from 'views/search/list/message/search-message-list-item-wrapper';
 
 describe('SearchMessageListItemWrapper', () => {
