+++ src/views/search/parts/search-panel-header.tsx
@@ -18,7 +18,8 @@ import { t } from '@zextras/carbonio-shell-ui';
 import { useNavigate } from 'react-router-dom';
 
 import { SEARCH_ROUTE } from '../../../constants';
-import type { MailMessage, NormalizedConversation } from '../../../types';
+import { NormalizedConversation } from 'types/conversations';
+import { MailMessage } from 'types/messages';
 
 export const SearchPanelHeader: FC<{
 	item: NormalizedConversation | (Partial<MailMessage> & Pick<MailMessage, 'id'>);
